requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
camelot-py==1.0.0
certifi==2025.7.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
Deprecated==1.2.18
distro==1.9.0
et-xmlfile==2.0.0
fpdf2==2.7.9
h11==0.16.0
httpcore==1.0.9
httpx==0.28.0
idna==3.10
img2pdf==0.6.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
jiter==0.10.0
keyring==25.6.0
lxml==6.0.0
more-itertools==10.7.0
numpy==2.2.6
openai==1.14.0
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pikepdf==9.10.1
Pillow==11.3.0
pip==25.1.1
pycparser==2.22
pydantic<3,>=1.9.0
pydantic_core==2.33.2
pypdf==5.8.0
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-pptx==0.6.23
pytz==2025.2
pywin32-ctypes==0.2.3
Scrapy==2.11.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
tabula-py==2.10.0
tabulate==0.9.0
tqdm==4.67.1
typing_extensions>=4.9.0
typing-inspection>=0.4.0
tzdata==2025.2
wrapt==1.17.2
XlsxWriter==3.2.5
markdown-it-py==3.0.0
openai==1.14.0
requests==2.31.0